+++ src/control.entity.ts
@@ -143,6 +143,13 @@ export function addControlEntity(
     }
   }
 
+  // Configure the vacuum.
+  // prettier-ignore
+  if (domain === 'vacuum') {
+    log.debug(`= vacuum device ${CYAN}${entity.entity_id}${db} activity: ${CYAN}${state.attributes['activity']}${db}`);
+    mutableDevice.addVacuum(endpointName);
+  }
+
   // Add command handlers
   for (const hassCommand of hassCommandConverter.filter((c) => c.domain === domain)) {
     log.debug(`- command: ${CYAN}${hassCommand.command}${db}`);